Create Data Processor class
Add implementation and definition for the abstract data processor class.

Make a correction in the DataProcessor.cpp file

Add a correction to the intializatio of the constructor of the class.

Make final edits before review

Removed process() from cpp file

Make a correction in the DataProcessor.cpp file

Add a correction to the intializatio of the constructor of the class.