Create the control_panel_creation class and the control_panel_config_full file.
The control panel is entirely created from the config file.
The control_panel_creation class reads through every JSON object and creates the corresponding widget. It then creates a dictionary of tab names and the widgets that are a part of it.
The process of creating the control panel is entirely dynamic, so all that needs to be updated is the config file and/or the widget objects